Unison.Util.Rope

Documentation

chunks :: Rope a -> [a]

singleton :: Sized a => a -> Rope a

one :: Sized a => a -> Rope a

map :: Sized b => (a -> b) -> Rope a -> Rope b

traverse :: (Applicative f, Sized b) => (a -> f b) -> Rope a -> f (Rope b)

null :: Sized a => Rope a -> Bool

flatten :: Monoid a => Rope a -> a

two :: Sized a => Rope a -> Rope a -> Rope a

cons :: (Sized a, Semigroup a) => a -> Rope a -> Rope a

uncons :: Sized a => Rope a -> Maybe (a, Rope a)

snoc :: (Sized a, Semigroup a) => Rope a -> a -> Rope a

unsnoc :: Sized a => Rope a -> Maybe (Rope a, a)

index :: (Sized a, Index a ch) => Int -> Rope a -> Maybe ch

debugDepth :: Rope a -> Int

class Sized a where

Methods

size :: a -> Int

class Take a where

Methods

take :: Int -> a -> a

class Drop a where

Methods

drop :: Int -> a -> a

class Reverse a where

Methods

reverse :: a -> a

class Index a elem where

Methods

unsafeIndex :: Int -> a -> elem

data Rope a

Roughly size-balanced binary tree of chunks. There are a few operations that are sloppier about rebalancing as long as that can't lead to trees of more than logarithmic depth.

The Int in the Two constructor is a cached size of that subtree.